Add files via upload
The transmission of electrical energy from source to load for a distance without any conducting wire or cables is called Wireless Power Transmission. The concept of wireless power transfer was realized by Nikola Tesla. Wireless power transfer can make a remarkable change in the field of the electrical engineering which eliminates the use conventional copper cables and current carrying wires.
Day by day new technologies are making our life simpler. Wireless charging through resonance could be one of the next technologies that bring the future nearer. In this project it has been shown that it is possible to charge low power devices wirelessly.
It minimizes the complexity that arises for the use of conventional wire system. In addition, the project also opens up new possibilities of wireless systems in our other daily life uses.
The objective of this project is to design and construct a method to transmit wireless electrical power through space and charge a designated low power device.
Wireless power transfer will be next biggest move in this integrated technologies world because it has numerous advantages and applications.

Advantages : 
   1. Complete removal of wired connections is possible by Wireless Power Transform(WPT).
   2. User friendly & reduces complexity.
   
Disadvantages:
1. Major disadvantage of wireless power transfer is high power loss.
2. Not suitable for power transfer for long distances.